backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
pymysql==1.1.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic-settings==2.6.1
python-multipart==0.0.12
email-validator>=2.0.0anthropic>=0.28.0